requirements.txt
gunicorn
firebase_admin
flask
Flask-Compress
python-dotenv
htmlmin